:truck: refactor(architecture): restructure folder structure
- Reorganize and relocate files related to Prisma client and Prisma repository as part of the restructuring
- Update Prisma client imports in the affected files
